+++ game/src/client/Boomberman.h
@@ -26,7 +26,7 @@ public:
   void setBoombermanPos(int new_x, int new_y);
   bool move(Map *map, std::shared_ptr<int[]> curr_pos, int x, int y);
   void cleanUp();
-  void gotHit(double long when);
+  void gotHit(int64_t when);
   void decrementIframes();
   void animateHit();
   int getState() const;
